moved gyro to layer image
